+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Solve bugs with opening urls on certain platforms.
 * Try to solve issues with creating pull requests on some platforms.
+* Handle creating pull request when on master branch.
+* Properly handle creation of pull request if access token is not yet set.
+* Fix small bugs.
 
 ## 0.0.3